video/requirements.txt
google-cloud-automl >= 0.9.0